Extract directory constants to utils
