+++ src/components/structures/RoomDirectory.tsx
@@ -28,7 +28,7 @@ import { _t } from '../../languageHandler';
 import SdkConfig from '../../SdkConfig';
 import { instanceForInstanceId, protocolNameForInstanceId } from '../../utils/DirectoryUtils';
 import Analytics from '../../Analytics';
-import NetworkDropdown, { ALL_ROOMS, Protocols } from "../views/directory/NetworkDropdown";
+import NetworkDropdown from "../views/directory/NetworkDropdown";
 import SettingsStore from "../../settings/SettingsStore";
 import { IDialogProps } from "../views/dialogs/IDialogProps";
 import AccessibleButton, { ButtonEvent } from "../views/elements/AccessibleButton";
@@ -43,6 +43,7 @@ import PosthogTrackers from "../../PosthogTrackers";
 import { PublicRoomTile } from "../views/rooms/PublicRoomTile";
 import { getFieldsForThirdPartyLocation, joinRoomByAlias, showRoom } from "../../utils/rooms";
 import { GenericError } from "../../utils/error";
+import { ALL_ROOMS, Protocols } from "../../utils/DirectoryUtils";
 
 const LAST_SERVER_KEY = "mx_last_room_directory_server";
 const LAST_INSTANCE_KEY = "mx_last_room_directory_instance";
